Extract handle sumbit answers in a new file
